% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getEIC}
\alias{getEIC}
\title{Recalculate the EIC}
\usage{
getEIC(clevCov, pred_star, n)
}
\arguments{
\item{clevCov}{...}

\item{pred_star}{...}

\item{n}{...}
}
\description{
Recompute the EIF based on the previously calculated clever covariates.
}
