% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrEstimate.R
\name{hdrEst}
\alias{hdrEst}
\title{h-density estimation}
\usage{
hdrEst(fit, s, i, B, t, Anode, intervention = 1, MC = 1000)
}
\arguments{
\item{fit}{\code{fit} object obtained by \code{initEst}.}

\item{s}{Where we are in the s loop (part of the clever covariate calculation).}

\item{i}{Where we are in the i loop (part of the clever covariate calculation).}

\item{B}{Number of observations to sample from P and P^*.}

\item{t}{Outcome time point of interest. It must be greater than the intervention node A.}

\item{Anode}{Intervention node.}

\item{intervention}{Specify g^*, of P(A|past). Right now supports only 1/0 type interventions.}

\item{MC}{How many Monte Carlo samples should be generated.}
}
\description{
This function evaluates the h-density ratios, necessary for the clever covariate calculation. 
It approximates the ratio of densities under intervened and non-intervened distributions.
}
