% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_H_densities.R
\name{hstarEst}
\alias{hstarEst}
\title{h*-density estimation}
\usage{
hstarEst(fit, s, i, B, t)
}
\arguments{
\item{fit}{\code{fit} object obtained by \code{initEst}.}

\item{s}{Where we are in the s loop (part of the clever covariate calculation).}

\item{i}{Where we are in the i loop (part of the clever covariate calculation).}

\item{B}{Number of observations to sample from P and P^*.}

\item{t}{Outcome time point of interest. It must be greater than the intervention node A.}
}
\value{
An object of class \code{tstmle01}.
\describe{
\item{h_cy_star}{Empirical estimate of h_cy under the intervention g^*.}
\item{h_ca_star}{Empirical estimate of h_ca under the intervention g^*.}
\item{h_cw_star}{Empirical estimate of h_cw under the intervention g^*.}}
}
\description{
This function evaluates the %h^*-density, necessary for the clever covariate calculation.
}
