% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmleEstimate.R
\name{mainTMLE}
\alias{mainTMLE}
\title{Main TMLE Calculations}
\usage{
mainTMLE(fit, t, Anode, MC, maxIter = 50, tol = 10^-5)
}
\arguments{
\item{fit}{\code{fit} object obtained by \code{initEst}.}

\item{t}{Outcome time point of interest. It must be greater than the intervention node A.}

\item{Anode}{Intervention node.}

\item{MC}{How many Monte Carlo samples should be generated.}

\item{maxIter}{Maximum number of iterations.}

\item{tol}{Lower bound for epsilon.}
}
\value{
An object of class \code{tstmle}.
\describe{
\item{psi}{Estimate of the target parameter.}
\item{var.psi}{Variance, based on the influence function.}}
}
\description{
This function runs the TMLE and computes the estimate.
}
