% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matchProb}
\alias{matchProb}
\title{Match probability}
\usage{
matchProb(fit, node, covs)
}
\arguments{
\item{fit}{\code{fit} object obtained by \code{initEst}.}

\item{node}{node for which we want the probability of success based on observed covariates}

\item{covs}{set of covariate observations, in the correct order}
}
\description{
Match probabilities of success based on covariate info
}
