% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdunif}
\alias{rdunif}
\title{Discrete Uniform Distribution}
\usage{
rdunif(n, k)
}
\arguments{
\item{n}{Number of points to sample from.}

\item{k}{Number of samples sampled.}
}
\description{
Sample from discrete uniform distribution
}
