% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_samp_data.R
\docType{data}
\name{ts_samp_data_ex}
\alias{ts_samp_data_ex}
\title{Shorter Mock Binary Time Series Data Set}
\format{A \code{data.frame} with 3 columns.
\describe{
  \item{Y}{A binary variable representing an outcome of interest.}
  \item{A}{A binary variable representing an intervention of interest.}
  \item{W}{A binary variable representing a baseline covariate of interest.}
}}
\usage{
ts_samp_data_ex
}
\description{
A dataset with a simple data structure O = (W, A, Y), where all covariates in
the data set are binary (0/1). This is a simple dataset designed specifically
to illustrate the TMLE estimation procedure.
}
\keyword{datasets}
