% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverCovEstimate.R
\name{cleverCov}
\alias{cleverCov}
\title{Clever Covariate calculation}
\usage{
cleverCov(fit, t, Anode, intervention = 1, MC = 1000, B = 100)
}
\arguments{
\item{fit}{\code{fit} object obtained by \code{initEst}.}

\item{t}{Outcome time point of interest. It must be greater than the intervention node A.}

\item{Anode}{Intervention node.}

\item{intervention}{Specify g^*, of P(A|past). Right now supports only 1/0 type interventions.}

\item{MC}{How many Monte Carlo samples should be generated.}

\item{B}{How many samples to draw from P and P^*, as part of the h-density estimation.}
}
\value{
An object of class \code{tstmle}.
\describe{
\item{fit}{Fit objects for each part of the likelihood.}}
}
\description{
This function calculates the clever covariate for each time point.
}
