% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getEIC}
\alias{getEIC}
\title{Recalculate the EIC}
\usage{
getEIC(clevCov, pred_star, n)
}
\arguments{
\item{clevCov}{\code{clevCov} object obtained by \code{cleverCov}.}

\item{pred_star}{Updated predictions from \code{mainTMLE}.}

\item{n}{Number of samples (time-points).}
}
\description{
Recompute the EIF based on the previously calculated clever covariates.
}
