% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPred}
\alias{getPred}
\title{Get Predictions}
\usage{
getPred(fit, t)
}
\arguments{
\item{fit}{\code{fit} object obtained by \code{initEst}.}

\item{t}{Outcome time point of interest. It must be greater than the intervention node A.}
}
\description{
Get actual and predicted values for time t, based on the previously obtained
fit and actual Cs. Minimum t is 1.
}
