% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hEstimate.R
\name{hEst}
\alias{hEst}
\title{h-density estimation}
\usage{
hEst(fit, i, B, t, Anode)
}
\arguments{
\item{fit}{\code{fit} object obtained by \code{initEst}.}

\item{i}{Where we are in the i loop (part of the clever covariate calculation).}

\item{B}{Number of observations to sample from P and P^*.}

\item{t}{Outcome time point of interest. It must be greater than the intervention node A.}

\item{Anode}{Intervention node.}
}
\description{
This function evaluates the h-density, necessary for the clever covariate calculation.
}
