% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initEstimate.R
\name{initEst}
\alias{initEst}
\title{Initial Estimate}
\usage{
initEst(data, freqW = NULL, freqA = NULL, freqY = NULL)
}
\arguments{
\item{data}{data.frame object containg the time series with relevant time ordering.}

\item{freqW}{A numeric specifying the Markov order for W nodes.}

\item{freqA}{A numeric specifying the Markov order for A nodes.}

\item{freqY}{A numeric specifying the Markov order for Y nodes.}
}
\description{
This function generates coefficients of previous variables based on a single time series observation.
Later on, this should include more complex relationships instead of just past variables and 
better estimation methods.
}
