% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmleEstimate.R
\name{mainTMLE}
\alias{mainTMLE}
\title{Main TMLE Calculations}
\usage{
mainTMLE(clevCov, fit, maxIter = 50, tol = 10^-5)
}
\arguments{
\item{clevCov}{Results object obtained by \code{cleverCov}. It contains clever covariates for each of 
the components of the likelihood, and the efficient influence curve.}

\item{fit}{\code{fit} object obtained by \code{initEst}.}

\item{maxIter}{Maximum number of iterations.}

\item{tol}{Lower bound for epsilon.}
}
\description{
This function runs the TMLE and computes the estimate.
}
