% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_tmle.R
\name{mainTMLE}
\alias{mainTMLE}
\title{Main TMLE Calculations}
\usage{
mainTMLE(fit, t, Anode, intervention = NULL, alpha = 0.05, B = 100,
  N = 100, MC = 100, maxIter = 50, tol = 10^-5)
}
\arguments{
\item{fit}{\code{fit} object obtained by \code{initEst}.}

\item{t}{Outcome time point of interest. It must be greater than the intervention node A.}

\item{Anode}{Intervention node.}

\item{intervention}{Specify %g^*, of %P(A \mid \text{past}). Right now, this supports only 1/0 type interventions.}

\item{alpha}{alpha}

\item{B}{How many samples to draw from P, as part of the h-density estimation.}

\item{N}{How many sample to draw from %P^*, as part of the h-density estimation.}

\item{MC}{How many Monte Carlo samples should be generated.}

\item{maxIter}{Maximum number of iterations.}

\item{tol}{Lower bound for epsilon.}
}
\value{
An object of class \code{tstmle}.
\describe{
\item{psi}{Estimate of the target parameter.}
\item{var.psi}{Variance, based on the influence function.}}
\item{IC}{Influence curve.}
}
\description{
This function runs the TMLE and computes the estimate.
}
