% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcEstimate.R
\name{mcEst}
\alias{mcEst}
\title{Compute the Monte Carlo estimate}
\usage{
mcEst(fit, start = 1, node = "W", t, Anode, intervention = NULL,
  lag = 0, MC, init = FALSE, returnMC = FALSE, clevCov = FALSE,
  set = NULL)
}
\arguments{
\item{fit}{\code{fit} object obtained by \code{initEst}.}

\item{start}{Start generating Monte Carlo estimates from this point O_i. The earliest time point is 1. For the clever covariate, this is also our s.}

\item{node}{Start generating Monte Carlo estimates from O_i node W,A or Y.}

\item{t}{Outcome time point of interest. It must be greater than the intervention node A.}

\item{Anode}{Intervention node.}

\item{intervention}{Specify g^*, of P(A|past). Right now supports only 1/0 type interventions.}

\item{lag}{This is an user impossed dependency lag necessary for the calculation of the clever covariate in the targeting step. 
It refers to O_i, and it is of the same dimension as the actual C(i), but further in the past.
Default is 0, meaning that it starts right after the node in question. Also, note that lag should be equal or smaller than start.}

\item{MC}{How many Monte Carlo samples should be generated.}

\item{init}{If TRUE, estimates the mean difference between the estimated Y_g*= 1 and 
Y_g* = 0 (Y under intervention). If FALSE, we are within a s loop and look at the Y before starting MC draws.}

\item{returnMC}{If TRUE, returns all MC draws.}

\item{clevCov}{If TRUE, this MC is used for the calculation of the clever covariate. Instead of observed data,
it used intervened P* for further MC draws.}

\item{set}{Set the s node to either 1 or 0. Used for the clever covariate calculation.}
}
\value{
An object of class \code{tstmle}.
\describe{
\item{estimate}{Mean of the outcome at time t under specified intervention, or no intervention.}
\item{outcome}{Outcome at time t for each MC interation.}
\item{intervention}{Intervention specified.}
\item{MC}{How many Monte Carlo samples should be generated.}
\item{Anode}{Intervention node as a function of O_i.}
\item{s1}{Mean of the intervened outcome given s=1 or s=0 (used for the clever covariate calculation).} 
\item{s_full}{Intervened outcomes given s=1 or s=0.}  
\item{MCdata}{If \code{returnMC} is TRUE, returns a data.frame with MC time-series.} 
}
}
\description{
This function computes the Monte Carlo estimate of the expected value of the parameter of interest
under specified intervention and under the estimated mechanism.
}
